Moving stream_recorded_llm_response to ai_caller_replay
